requirements.txt
Django==4.1.7
django-browser-reload==1.6.0
django-jazzmin==2.6.0
python-docx==0.8.11
